+++ middleware/system/bf0_common.c
@@ -287,9 +287,9 @@ int sys_init_lh_debug_queue(void)
 
     q_cfg.qid = SYS_LH_DEBUG_QUEUE;
     q_cfg.tx_buf_size = 0;
-    q_cfg.tx_buf_addr = NULL;
-    q_cfg.tx_buf_addr_alias = NULL;
-    q_cfg.rx_buf_addr = NULL;
+    q_cfg.tx_buf_addr = (uint32_t)NULL;
+    q_cfg.tx_buf_addr_alias = (uint32_t)NULL;
+    q_cfg.rx_buf_addr = (uint32_t)NULL;
     q_cfg.rx_ind = debug_queue_rx_ind;
     q_cfg.user_data = 0;
